// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  READER: 'READER',
  AUTHOR: 'AUTHOR',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const BookStatus = {
  ONGOING: 'ONGOING',
  COMPLETED: 'COMPLETED',
  HIATUS: 'HIATUS'
} as const

export type BookStatus = (typeof BookStatus)[keyof typeof BookStatus]


export const NotificationType = {
  NEW_CHAPTER: 'NEW_CHAPTER',
  SYSTEM: 'SYSTEM'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]